backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2
django-webpack-loader==0.7.0
docxcompose==1.3.3
docxtpl==0.14.2
et-xmlfile==1.1.0
Fiona==1.8.20
geojson==2.5.0
geopandas==0.10.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
munch==2.5.0
mysqlclient==2.0.3
numpy==1.21.2
openpyxl==3.0.9
packaging==20.9
pandas==1.3.3
Pillow==8.3.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyproj==3.2.1
pytest==6.2.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
requests==2.25.1
rply==0.7.8
service-identity==18.1.0
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
transliterate==1.10.2
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.3
zope.interface==5.4.0
